# .junie/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearerOrKey": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "email": "votre.email@example.com",
      "name": "Votre Nom"
    },
    "description": "Backend FastAPI/SQLite (compatible PostgreSQL) pour la gestion des utilisateurs, environnements, groupes et permissions.\n\nA service wich deploys e home web platform on a Raspberry Pi or a VPS\nThis is usable for debian, Ubuntu systems. other Platforms are planned\n\n## API Overview\nThis API provides 108 operations across 78 endpoints. Supports DELETE, GET, POST, PUT HTTP methods.\n\n## Authentication\nMost endpoints (93%) require authentication.\n\n## Data Formats\nSupports the following content types: application/json, application/x-www-form-urlencoded.\n\n## Response Codes\nSuccess responses: 200. Error responses: 400, 401, 403, 404, 409, 418, 422, 500, 503.\n\n## Data Models\nIncludes 31 data model definitions for request and response validation.\n\n## Error Handling\nThe API returns standardized error responses with appropriate HTTP status codes and detailed error messages.",
    "license": {
      "name": "Apache",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "App API",
    "version": "0.1.0",
    "x-generated-at": "2025-07-05T01:47:45.385528",
    "x-generator": "app OpenAPI Generator"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/environments/{environment_id}/users": {
      "get": {
        "description": "Renvoie tous les utilisateurs ayant accès à un environnement via une policy (via les rules).",
        "operationId": "get_environment_users_environments__environment_id__users_get",
        "parameters": [
          {
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "title": "Environment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Environment Users Environments  Environment Id  Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Utilisateurs récupérés avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Environnement non trouvé"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Utilisateurs liés à un environnement",
        "tags": [
          "environments"
        ]
      }
    },
...
    "/groups/{group_id}/users": {
      "get": {
        "description": "Récupère la liste de tous les utilisateurs appartenant à un groupe spécifique",
        "operationId": "list_group_users_groups__group_id__users_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Group Users Groups  Group Id  Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Liste des utilisateurs du groupe récupérée avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Groupe non trouvé"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Lister les utilisateurs d'un groupe",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{group_id}/users/{user_id}": {
      "delete": {
        "description": "Retire un utilisateur spécifique d'un groupe",
        "operationId": "remove_user_groups__group_id__users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove User Groups  Group Id  Users  User Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Utilisateur retiré du groupe avec succès ou n'était pas dans le groupe"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Groupe ou utilisateur non trouvé"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Retirer un utilisateur",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "description": "Ajoute un utilisateur spécifique à un groupe",
        "operationId": "assign_user_groups__group_id__users__user_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Assign User Groups  Group Id  Users  User Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Utilisateur ajouté au groupe avec succès ou déjà présent dans le groupe"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Groupe ou utilisateur non trouvé"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Associer un utilisateur",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/organizations/{org_id}/users": {
      "get": {
        "description": "Récupère tous les utilisateurs ayant accès à une organisation spécifique.",
        "operationId": "list_organization_users_organizations__org_id__users_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Organization Users Organizations  Org Id  Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Utilisateurs de l'organisation récupérés avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Organisation non trouvée"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Lister les utilisateurs d'une organisation",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/policies/{policy_id}/users": {
      "get": {
        "description": "Récupère la liste de tous les utilisateurs associés à une policy spécifique.",
        "operationId": "list_policy_users_policies__policy_id__users_get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Policy Users Policies  Policy Id  Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Liste des utilisateurs de la policy récupérée avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission insuffisante"
          },
          "404": {
            "description": "Policy non trouvée"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Lister les utilisateurs d'une policy",
        "tags": [
          "policies"
        ]
      }
    },
...
  },
  "security": [
    {
      "OAuth2PasswordBearerOrKey": []
    }
  ],
  "servers": [
    {
      "description": "Development server (local)",
      "url": "http://localhost:8010"
    }
  ],
...
}